+++ tests/SimpleTest.php
@@ -5,10 +5,8 @@
  * @package Spawn
  */
 
-echo ">>> SimpleTest.php LOADED <<<\n";
-
 /**
- * Simple test class - no WP dependencies.
+ * Simple test class.
  */
 class SimpleTest extends \PHPUnit\Framework\TestCase {
 
@@ -16,7 +14,6 @@ class SimpleTest extends \PHPUnit\Framework\TestCase {
 	 * Test that true is true.
 	 */
 	public function test_true_is_true(): void {
-		fwrite( STDERR, ">>> test_true_is_true RUNNING <<<\n" );
 		$this->assertTrue( true );
 	}
 
@@ -28,7 +25,7 @@ class SimpleTest extends \PHPUnit\Framework\TestCase {
 	}
 
 	/**
-	 * Test that Config class exists.
+	 * Test that Config class exists after WordPress loads.
 	 */
 	public function test_config_class_exists(): void {
 		$this->assertTrue( class_exists( 'Spawn\\Config' ) );